Module Sarek_cpu_runtime_pools.ThreadPool

type work_item = {

kernel : Sarek_cpu_runtime_types.thread_state -> Sarek_cpu_runtime_types.shared_mem -> Spoc_framework.Framework_sig.exec_arg array -> unit;

args : Spoc_framework.Framework_sig.exec_arg array;

uses_barriers : bool;
(*
Whether kernel uses barriers
*)

start_tid : int;
(*
First global thread ID (inclusive)
*)

end_tid : int;
(*
Last global thread ID (exclusive)
*)

start_block : int;
(*
First block ID (inclusive)
*)

end_block : int;
(*
Last block ID (exclusive)
*)

bx : int;

by : int;

bz : int;

gx : int;

gy : int;

gz : int;

}

Work item: either thread or block distribution based on barrier usage.

No barriers: thread distribution (start_tid/end_tid) - more granular

With barriers: block distribution (start_block/end_block) - preserves BSP

type t = {

num_workers : int;

mutable workers : unit Stdlib.Domain.t array;

work : work_item option Stdlib.Atomic.t array;

mutex : Stdlib.Mutex.t;

work_ready : Stdlib.Condition.t;

work_done : Stdlib.Condition.t;

mutable pending_workers : int;

mutable shutdown : bool;

mutable generation : int;

}

val run_block_bsp : block:(int * int * int) -> grid:(int * int * int) -> block_idx:(int * int * int) -> (Sarek_cpu_runtime_types.thread_state -> Sarek_cpu_runtime_types.shared_mem -> Spoc_framework.Framework_sig.exec_arg array -> unit) -> Spoc_framework.Framework_sig.exec_arg array -> unit

Run a single block with BSP barriers using Effect.Shallow fibers

val run_threads : work_item -> unit

Run a range of global threads directly - for barrier-free kernels. More efficient than block distribution when no barriers are needed.

val worker_fn : t -> int -> unit

Worker function - runs in each domain, processes work items

val create : int -> t

Create thread pool with N workers

val run_kernel : t -> block:(int * int * int) -> grid:(int * int * int) -> uses_barriers:bool -> (Sarek_cpu_runtime_types.thread_state -> Sarek_cpu_runtime_types.shared_mem -> Spoc_framework.Framework_sig.exec_arg array -> unit) -> Spoc_framework.Framework_sig.exec_arg array -> unit

Submit kernel to thread pool and wait for completion. Distribution strategy depends on barrier usage:

uses_barriers=false: distribute threads (more granular)

uses_barriers=true: distribute blocks (preserves BSP)

val shutdown : t -> unit