Module type Memory.BUFFER

A buffer packages backend-specific buffer with its operations. All transfers use raw pointers with byte sizes to avoid type parameter escaping issues in first-class modules.

val device : Device.t

The device this buffer is allocated on

val size : int

Number of elements

val elem_size : int

Size of each element in bytes

val device_ptr : nativeint

Get raw device pointer (for kernel arg binding)

val host_ptr_to_device : unit Ctypes.ptr -> byte_size:int -> unit

Transfer from host pointer to device

val device_to_host_ptr : unit Ctypes.ptr -> byte_size:int -> unit

Transfer from device to host pointer

val bind_to_kargs : Spoc_framework.Framework_sig.kargs -> int -> unit

Bind this buffer to kernel args at given index

val free : unit -> unit

Free the buffer