Module Spoc_core.Vector_types

Element Types

type (_, _) scalar_kind =

| Float32 : (float, Stdlib.Bigarray.float32_elt) scalar_kind

| Float64 : (float, Stdlib.Bigarray.float64_elt) scalar_kind

| Int32 : (int32, Stdlib.Bigarray.int32_elt) scalar_kind

| Int64 : (int64, Stdlib.Bigarray.int64_elt) scalar_kind

| Char : (char, Stdlib.Bigarray.int8_unsigned_elt) scalar_kind

| Complex32 : (Stdlib.Complex.t, Stdlib.Bigarray.complex32_elt) scalar_kind

Standard numeric kinds backed by Bigarray

type 'a custom_type = {

elem_size : int;
(*
Size of each element in bytes
*)

get : unit Ctypes.ptr -> int -> 'a;
(*
Read element at index
*)

set : unit Ctypes.ptr -> int -> 'a -> unit;
(*
Write element at index
*)

name : string;
(*
Type name for debugging
*)

}

Custom type descriptor for ctypes-based structures

module Custom_helpers : sig ... end

Helper functions for custom type implementations. These wrap Ctypes operations to provide simpler APIs for PPX-generated code.

type (_, _) kind =

| Scalar : ('a, 'b) scalar_kind -> ('a, 'b) kind

| Custom : 'a custom_type -> ('a, unit) kind

Unified kind type supporting both scalar and custom types

Kind Helpers

val to_bigarray_kind : 'a 'b. ('a, 'b) scalar_kind -> ('a, 'b) Stdlib.Bigarray.kind

Convert scalar kind to Bigarray.kind

val scalar_elem_size : 'a 'b. ('a, 'b) scalar_kind -> int

Element size in bytes

val elem_size : 'a 'b. ('a, 'b) kind -> int

val scalar_kind_name : 'a 'b. ('a, 'b) scalar_kind -> string

Kind name for debugging

val kind_name : 'a 'b. ('a, 'b) kind -> string

Host Storage (GADT)

type (_, _) host_storage =

| Bigarray_storage : ('a, 'b, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t -> ('a, 'b) host_storage

| Custom_storage : {

ptr : unit Ctypes.ptr;

custom : 'a custom_type;

length : int;

} -> ('a, unit) host_storage

CPU-side storage - either Bigarray or raw ctypes pointer

Location Tracking

type location =

| CPU
(*
Data only on host
*)

| GPU of Device.t
(*
Data only on specific device
*)

| Both of Device.t
(*
Synced on host and device
*)

| Stale_CPU of Device.t
(*
GPU is authoritative, CPU outdated
*)

| Stale_GPU of Device.t
(*
CPU is authoritative, GPU outdated
*)

Where the authoritative copy of data resides

Device Buffer Abstraction

module type DEVICE_BUFFER = Memory.BUFFER

Device buffer reuses Memory.BUFFER module type. We use the raw module type here (not the phantom-typed Memory.buffer) because the hashtable stores buffers for a single vector type, and the type safety comes from the Vector's ('a, 'b) t type parameter.

type device_buffer = (module DEVICE_BUFFER)

type device_buffers = (int, device_buffer) Stdlib.Hashtbl.t

Device buffer storage - maps device ID to buffer

Vector Type

type ('a, 'b) t = {

host : ('a, 'b) host_storage;

device_buffers : device_buffers;

length : int;

kind : ('a, 'b) kind;

mutable location : location;

mutable auto_sync : bool;
(*
Enable automatic CPU sync on get
*)

id : int;
(*
Unique vector ID for debugging
*)

}

High-level vector with location tracking