sarek index

Library sarek

The entry point of this library is the module: Sarek.

Library sarek.codegen

The entry point of this library is the module: Sarek_codegen.

Library sarek.core

The entry point of this library is the module: Spoc_core.

Library sarek.core_base

The entry point of this library is the module: Spoc_core_base.

Library sarek.core_js

The entry point of this library is the module: Spoc_core_js.

Library sarek.df64

The entry point of this library is the module: Sarek_df64.

Library sarek.execute

The entry point of this library is the module: Sarek_execute.

Library sarek.float64

The entry point of this library is the module: Sarek_float64.

Library sarek.framework.registry

The entry point of this library is the module: Spoc_framework_registry.

Library sarek.frontend

This library exposes the following toplevel modules:

Sarek_ast

Sarek_convergence

Sarek_core_primitives

Sarek_debug

Sarek_defunc

Sarek_env

Sarek_error

Sarek_float_literal

Sarek_ir_ppx

Sarek_lower_ir

Sarek_mono

Sarek_parse

Sarek_parse_helpers

Sarek_ppx_registry

Sarek_quote_ir

Sarek_reserved

Sarek_scheme

Sarek_tag_erasure

Sarek_tailrec

Sarek_tailrec_analysis

Sarek_tailrec_bounded

Sarek_tailrec_elim

Sarek_tailrec_pragma

Sarek_typed_ast

Sarek_typer

Sarek_types

Sarek_unsupported Refusal text for the OCaml constructs the Sarek kernel parser does not implement (backlog-192).

Library sarek.gemm

The entry point of this library is the module: Sarek_gemm.

Library sarek.geometry

The entry point of this library is the module: Sarek_geometry.

Library sarek.interp

The entry point of this library is the module: Sarek_interp.

Library sarek.native_gen

This library exposes the following toplevel modules:

Sarek_native_gen

Sarek_native_gen_base

Sarek_native_gen_expr

Sarek_native_gen_kernel

Sarek_native_helpers

Sarek_native_intrinsics

Sarek_quote

Library sarek.ppx

The entry point of this library is the module: Sarek_ppx.

Library sarek.ppx.intrinsic

The entry point of this library is the module: Sarek_ppx_intrinsic.

Library sarek.ppx.lib

The entry point of this library is the module: Sarek_ppx_lib.

Library sarek.real64

The entry point of this library is the module: Sarek_real64.

Library sarek.stdlib

The entry point of this library is the module: Sarek_stdlib.

Library sarek.stdlib_meta

The entry point of this library is the module: Sarek_stdlib_meta.

Library sarek.transpile

The entry point of this library is the module: Sarek_transpile.

Library sarek.tuple_vec

The entry point of this library is the module: Sarek_tuple_vec.

Library sarek.visibility

The entry point of this library is the module: Visibility_lib.

Library sarek.webgpu_runtime

The entry point of this library is the module: Spoc_webgpu_runtime.

Library sarek.worklist

The entry point of this library is the module: Sarek_worklist.