sarek index

Library sarek

The entry point of this library is the module: Sarek.

Library sarek.codegen

The entry point of this library is the module: Sarek_codegen.

Library sarek.core

The entry point of this library is the module: Spoc_core.

Library sarek.float64

The entry point of this library is the module: Sarek_float64.

Library sarek.framework.registry

The entry point of this library is the module: Spoc_framework_registry.

Library sarek.frontend

This library exposes the following toplevel modules:

Sarek_ast

Sarek_convergence

Sarek_core_primitives

Sarek_debug

Sarek_env

Sarek_error

Sarek_ir_ppx

Sarek_lower_ir

Sarek_mono

Sarek_parse

Sarek_parse_helpers

Sarek_ppx_registry

Sarek_quote_ir

Sarek_reserved

Sarek_scheme

Sarek_tailrec

Sarek_tailrec_analysis

Sarek_tailrec_bounded

Sarek_tailrec_elim

Sarek_tailrec_pragma

Sarek_typed_ast

Sarek_typer

Sarek_types

Library sarek.geometry

The entry point of this library is the module: Sarek_geometry.

Library sarek.native_gen

This library exposes the following toplevel modules:

Kirc_Ast

Sarek_lower

Sarek_native_gen

Sarek_native_gen_base

Sarek_native_gen_expr

Sarek_native_gen_kernel

Sarek_native_helpers

Sarek_native_intrinsics

Sarek_quote

Library sarek.ppx

The entry point of this library is the module: Sarek_ppx.

Library sarek.ppx.intrinsic

The entry point of this library is the module: Sarek_ppx_intrinsic.

Library sarek.ppx.lib

The entry point of this library is the module: Sarek_ppx_lib.

Library sarek.stdlib

The entry point of this library is the module: Sarek_stdlib.

Library sarek.visibility

The entry point of this library is the module: Visibility_lib.